Earlier fixed parameters can be set, defaulting to earlier fixed value.
The hl part sets the search term language.

The earlier error handling is improved, instead of stopifnot, an explicit stop is made with the error code in case of an error. It will not effect running but give a meaningful error message in case of an error, such as a malformed request.