The get_widget () was amended with explicit tz and hl parameters.
However, only the hl is invoked, which is passed on from the gtrends hl parameter. It will correctly set the query language, right now even if g_trends(, hl = "fr-FR") is invoked it switches to English.

The tz part is not used, and defaults to the same time zone as earlier. Some documentation of Google's tz codes would be needed, but I do not find it, to correctly change this.

This is still a problem, because in any time zone apart from UTC -9 the "last hour" request will return a different time frame.