The tz parameter does not appear to change results. However, it is unclear how it works. I set it in the code to the same tz=300 value.
The hl part is correctly passed on to the widget and it further passes it on interest over time, region, etc.

It was earlier set constantly to en-US, even though in the main gtrends() function the user was supposed to be able to set it to 'fr-FR' or 'hu-HU'.

I changed the request url from google.fr to google.com, again for consistency within zzz. I believe that google.fr will eventually redirect to google.com apart from France, but may have some confusion on returned metadata.