Update zzz.R
Changed the cookie URL to be at the top of the document so it can be easily changed in the future. Also changed it to the trends.google.com alternative cookie. Also allows the user to replace it manually, e.g., by gtrendsR:::.pkgenv$cookie_url<-'the url'. An alternative would be to put it inside one of the top-level functions as an argument.