Update gtrends.R
Added "cookie_url" parameter, so that the user can use a different one of s/he chooses. Also means we can simply update the default in this file in the future if it changes. The trends.google.come NID cookie seems preferable to the previous version, which is blocked by firewalls that don't allow social media. Also, it is what is used by https://github.com/GeneralMills/pytrends.