Mixing topic and search keywords
There is a small problem when mixing search terms and topic ids. In that the payload2 in the interest_over_time()-function cannot be built since the returned widget from the get_widget()-function is off by one unit. I have fixed it in the zzz.R file temporarily. Perhaps there is a better way to deal with it. I added a reproducible example.